+++ d/phobos2/std/intrinsic.d
@@ -5,19 +5,19 @@
 
 /** These functions are built-in intrinsics to the compiler.
  *
-	Intrinsic functions are functions built in to the compiler,
-	usually to take advantage of specific CPU features that
-	are inefficient to handle via external functions.
-	The compiler's optimizer and code generator are fully
-	integrated in with intrinsic functions, bringing to bear
-	their full power on them.
-	This can result in some surprising speedups.
+        Intrinsic functions are functions built in to the compiler,
+        usually to take advantage of specific CPU features that
+        are inefficient to handle via external functions.
+        The compiler's optimizer and code generator are fully
+        integrated in with intrinsic functions, bringing to bear
+        their full power on them.
+        This can result in some surprising speedups.
  *
  * Copyright: Public Domain
  * License:   Public Domain
  * Authors:   Walter Bright
  * Macros:
- *	WIKI=Phobos/StdIntrinsic
+ *      WIKI=Phobos/StdIntrinsic
  */
 
 /* NOTE: This file has been patched from the original DMD distribution to
@@ -38,25 +38,25 @@ nothrow:
  * Scans the bits in v starting with bit 0, looking
  * for the first set bit.
  * Returns:
- *	The bit number of the first bit set.
- *	The return value is undefined if v is zero.
+ *      The bit number of the first bit set.
+ *      The return value is undefined if v is zero.
  */
-pure nothrow int bsf(uint v);
+pure nothrow int bsf(size_t v);
 
 /**
  * Scans the bits in v from the most significant bit
  * to the least significant bit, looking
  * for the first set bit.
  * Returns:
- *	The bit number of the first bit set.
- *	The return value is undefined if v is zero.
+ *      The bit number of the first bit set.
+ *      The return value is undefined if v is zero.
  * Example:
  * ---
  * import std.stdio;
  * import std.intrinsic;
  *
  * int main()
- * {   
+ * {
  *     uint v;
  *     int x;
  *
@@ -66,7 +66,7 @@ pure nothrow int bsf(uint v);
  *     x = bsr(v);
  *     writefln("bsr(x%x) = %d", v, x);
  *     return 0;
- * } 
+ * }
  * ---
  * Output:
  *  bsf(x21) = 0<br>
@@ -77,7 +77,7 @@ pure nothrow int bsr(uint v);
 /**
  * Tests the bit.
  */
-pure nothrow int bt(in uint *p, uint bitnum);
+pure nothrow int bt(in uint *p, size_t bitnum);
 
 /**
  * Tests and complements the bit.
@@ -87,7 +87,7 @@ int btc(uint *p, uint bitnum);
 /**
  * Tests and resets (sets to 0) the bit.
  */
-int btr(uint *p, uint bitnum);
+int btr(uint *p, size_t bitnum);
 
 /**
  * Tests and sets the bit.
@@ -99,16 +99,16 @@ int btr(uint *p, uint bitnum);
 p[index / (uint.sizeof*8)] & (1 << (index & ((uint.sizeof*8) - 1)))
 ---
  * Returns:
- * 	A non-zero value if the bit was set, and a zero
- *	if it was clear.
+ *      A non-zero value if the bit was set, and a zero
+ *      if it was clear.
  *
- * Example: 
+ * Example:
  * ---
 import std.stdio;
 import std.intrinsic;
 
 int main()
-{   
+{
     uint array[2];
 
     array[0] = 2;
@@ -130,7 +130,7 @@ int main()
     writefln("array = [0]:x%x, [1]:x%x", array[0], array[1]);
 
     return 0;
-} 
+}
  * ---
  * Output:
 <pre>
@@ -146,13 +146,13 @@ bt(array, 1) = -1
 array = [0]:x2, [1]:x100
 </pre>
  */
-int bts(uint *p, uint bitnum);
+int bts(uint *p, size_t bitnum);
 
 
 /**
  * Swaps bytes in a 4 byte uint end-to-end, i.e. byte 0 becomes
-	byte 3, byte 1 becomes byte 2, byte 2 becomes byte 1, byte 3
-	becomes byte 0.
+        byte 3, byte 1 becomes byte 2, byte 2 becomes byte 1, byte 3
+        becomes byte 0.
  */
 pure uint bswap(uint v);
